feat(transactions): implement transaction creation

+++ database/migrations/2025_09_20_193254_create_transactions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('amount');
+            $table->decimal('amount', 10, 2);
             $table->enum('type', ['Receita', 'Despesa']);
             $table->date('date');
             $table->string('description');
